gym_backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.35
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic[email]==2.6.1
python-multipart==0.0.9
